+++ packages/gamut/src/List/List.tsx
@@ -29,6 +29,11 @@ export interface ListProps extends AllListProps<ComponentProps<typeof ListEl>> {
    * How the List container should handle overflow.
    */
   overflow?: BoxProps['overflow'];
+  /**
+   * This is an override for the width of the wrapper element that contains the List.
+   * It is useful for custom scroll and breakpoint handling. Use with caution.
+   */
+  wrapperWidth?: BoxProps['width'];
 }
 
 export const List = forwardRef<HTMLUListElement, ListProps>(
@@ -49,6 +54,7 @@ export const List = forwardRef<HTMLUListElement, ListProps>(
       emptyMessage,
       overflow = 'auto',
       scrollToTopOnUpdate = false,
+      wrapperWidth,
       ...rest
     },
     ref
@@ -140,12 +146,12 @@ export const List = forwardRef<HTMLUListElement, ListProps>(
             as={isTable && !isEmpty && !loading ? 'table' : 'div'}
             data-testid={`scrollable-${id}`}
             height={isEmpty ? height : 'fit-content'}
-            maxWidth={1}
+            maxWidth={wrapperWidth || 1}
             minHeight={minHeight}
             overflow="inherit"
             position="relative"
             ref={!isEmpty ? tableRef : undefined}
-            width="inherit"
+            width={wrapperWidth || 'inherit'}
           >
             {content}
           </Box>