+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.530](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.529...@codecademy/gamut-kit@0.6.530) (2025-08-07)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.529](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.528...@codecademy/gamut-kit@0.6.529) (2025-08-06)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.528](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.527...@codecademy/gamut-kit@0.6.528) (2025-07-25)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.527](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.526...@codecademy/gamut-kit@0.6.527) (2025-07-25)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.526](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.525...@codecademy/gamut-kit@0.6.526) (2025-07-24)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
 ### [0.6.525](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.524...@codecademy/gamut-kit@0.6.525) (2025-07-22)
 
 **Note:** Version bump only for package @codecademy/gamut-kit